// src/modules/scheduling/enums/session-status.enum.ts
// src/scheduling/enums/session-status.enum.ts
export enum SessionStatus {
  PENDING_TUTOR_CONFIRMATION = 'PENDING_TUTOR_CONFIRMATION', // NUEVO
  SCHEDULED = 'SCHEDULED',
  PENDING_MODIFICATION = 'PENDING_MODIFICATION',
  REJECTED_BY_TUTOR = 'REJECTED_BY_TUTOR', // NUEVO
  CANCELLED_BY_STUDENT = 'CANCELLED_BY_STUDENT',
  CANCELLED_BY_TUTOR = 'CANCELLED_BY_TUTOR',
  CANCELLED_BY_ADMIN = 'CANCELLED_BY_ADMIN',
  COMPLETED = 'COMPLETED',
}